requirements.txt
requests>=2.31.0
httpx>=0.27.0
pytz>=2024.1
pyyaml>=6.0
playwright>=1.40.0
python-telegram-bot>=21.0
